resolve flake8 and move to core
